+++ Changes.md
@@ -11,6 +11,9 @@ Features
   - May be disabled entirely with `GafferScene.SceneAlgo.deregisterRenderAdaptor( "USDPointInstancerAdaptor" )`.
 - Viewer : Added "Expand USD Instancers" item to the Expansion menu. Defaults to on for all renderers except OpenGL.
 - PromotePointInstances : Added a new node for selectively converting a subset of a USD PointInstancer to expanded "hero" geometry.
+- Annotations :
+  - Added copy and paste of annotations. The right-click menu of an annotation allows you to copy the annotation. Pressing <kbd>Control</kbd> + <kbd>V</kbd> in the Node Editor will paste the annotation to the selected nodes.
+  - Double clicking on an annotation now pops up the annotation editor dialogue.
 
 Improvements
 ------------
@@ -43,6 +46,10 @@ API
 
 - EditScopeAlgo : Added `renameRenderPass()` and `renameRenderPassNonEditableReason()` functions.
 - SceneAlgo : Added `parallelGatherLocations()` function.
+- GraphGadget : Added `annotationsGadget()` function.
+- MetadataAlgo : Added `annotations()` variant accepting `Gaffer::Metadata::RegistrationTypes`. The default is `All` to match existing behavior and the previous `annotations()` variant is deprecated.
+- AnnotationsGadget : Added `annotationAt()` function.
+- AnnotationUI : Added `contextMenuSignal()` allowing customisations to the context menu for annotations.
 
 1.5.4.1 (relative to 1.5.4.0)
 =======